Move expiring event to update certifcates task
