+++ src/main/java/com/czertainly/core/tasks/SystemScheduledJobs.java
@@ -24,7 +24,6 @@ public class SystemScheduledJobs {
     public Void registerJobs() throws SchedulerException {
         schedulerService.registerScheduledJob(UpdateCertificateStatusTask.class);
         schedulerService.registerScheduledJob(UpdateIntuneRevocationRequestsTask.class);
-        schedulerService.registerScheduledJob(HandleExpiringCertificatesTask.class);
         return null;
     }
 }
